switch factory to produce a generator, which produces a ValidatingVisitor, to avoid issues where stale classes get reused, especially in tests
