+++ liquibase-standard/pom.xml
@@ -342,7 +342,7 @@
                         <param>liquibase.changeset.ChangeSetService</param>
                         <param>liquibase.parser.LiquibaseSqlParser</param>
                         <param>liquibase.database.LiquibaseTableNames</param>
-                        <param>liquibase.changelog.visitor.ValidatingVisitor</param>
+                        <param>liquibase.changelog.visitor.ValidatingVisitorGenerator</param>
                         <param>liquibase.io.OutputFileHandler</param>
                     </services>
                 </configuration>
