+++ liquibase-standard/src/main/java/liquibase/changelog/visitor/StandardValidatingVisitorGenerator.java
@@ -1,0 +1,13 @@
+package liquibase.changelog.visitor;
+
+public class StandardValidatingVisitorGenerator implements ValidatingVisitorGenerator {
+    @Override
+    public int getPriority() {
+        return PRIORITY_DEFAULT;
+    }
+
+    @Override
+    public ValidatingVisitor generateValidatingVisitor() {
+        return new ValidatingVisitor();
+    }
+}
